Modify some fun tidbits
